gemini_requirement.txt
google-genai
fastapi
dotenv
uvicorn